voice/Azure/requirements.txt
flask
openai
azure-cognitiveservices-speech
python-dotenv
websockets
asyncio
uuid
Quart
aiohttp
cors
OpenAIClient
logging
os